refactor(desktop): rename UI components to kebab-case
