+++ apps/desktop/src/renderer/onboarding-hero.tsx
