+++ apps/desktop/src/renderer/settings/providers-panel.tsx
